pyproject.toml
[project]
name = "reachy-vision-api"
version = "0.0.0"
description = " Lightweight YOLOv8 vision API for real-time object detection on Reachy robots."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi",
  "uvicorn",
  "ultralytics",
  "pillow",
  "python-multipart"
]

[project.optional-dependencies]
dev = [
  "ruff",
  "mypy",
  "pytest",
  "pytest-cov",
  "httpx",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "if __name__ == .__main__.:",
]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.0.0"
version_files = ["pyproject.toml:^version"]
tag_format = "v$version"
update_changelog_on_bump = true
